Move request logging tools to Gateway
